Extract 'Form2' to its own file
So it can be explicitly required by poste without requiring the whole
pegasus/src/database file
